requirements.txt
Django>=1.7
South==1.0.0
psycopg2==2.5.1
django-piston==0.2.3
pytz==2011n
chardet==1.0.1
python-dateutil==1.5
django-cliauth==0.9.1
django-mobile==0.3.0
django-haystack==2.1
pysolr==2.1
poster==0.8.1
etiquetando==0.1
html2text==3.200.3
django-taggit==0.12.1
python-memcached==1.53
django-hitcounter==0.1.1
Pillow==2.5.1
django-i18n-model==0.0.7
django-tastypie==0.11.0
gunicorn==19.1.0
eventlet==0.15.0
PyYAML==3.11

# Deps for sentry client (raven)
raven==3.5.2
tornado==3.1.1

# Deps for Single SignOn (SSO) - Replaced with django-browserid==0.9
django-browserid==0.11
django-revproxy==0.2.8

# Converse.js (XMPP client)
django-conversejs==0.3.1

# Feedzilla (planet) and deps
feedzilla==0.24
django-common==0.1.51
feedparser==5.1.3
lxml==3.2.4
grab==0.4.13
transliterate==1.5

# Diazo
diazo==1.0.5

# Dpaste
dpaste==2.8

# Mailman 2 REST API
mailman-api==0.2.1
